Allow role to be included multiple times
This role basically only calls `set_fact`, so it should be safe and
performant to include multiple times. While it is not expected that one
would explicitly include this role multiple times, it is possible that
multiple roles may depend on this role, and therefore duplicates should
be supported if one of these parent roles is included conditionally.

Take the following example of `roles` section in a playbook where both
`role.a` and `role.b` depend on this role:

roles:
  - role: role.a
    when: condition_that_is_false
  - role: role.b

As of ansible-core 2.15.0, this role wouldn't execute *at all* in the
above example because the first inclusion is skipped and the second one
isn't evaluated. Arguably this behavior should have been the case in
previous ansible-core releases, but it wasn't enforced as such.